+++ doc-examples/src/main/java/sting/doc/examples/autofragments/EntitiesAutoFragment.java
@@ -1,8 +1,0 @@
-package sting.doc.examples.autofragments;
-
-import sting.AutoFragment;
-
-@AutoFragment( "entities" )
-public interface EntitiesAutoFragment
-{
-}