+++ doc-examples/src/main/java/sting/doc/examples/autofragments/MessageRepository.java
@@ -1,7 +1,0 @@
-package sting.doc.examples.autofragments;
-
-public class MessageRepository
-{
-  //DOC ELIDE START
-  //DOC ELIDE END
-}