+++ processor/src/main/java/sting/processor/Registry.java
@@ -2,13 +2,10 @@ package sting.processor;
 
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
@@ -34,17 +31,6 @@ final class Registry
    */
   @Nonnull
   private final List<InjectorDescriptor> _injectors = new ArrayList<>();
-  /**
-   * The set of auto-fragments registered.
-   */
-  @Nonnull
-  private final Map<String, AutoFragmentDescriptor> _autoFragments = new HashMap<>();
-  /**
-   * The set of auto-fragments contributors registered.
-   */
-  @Nonnull
-  private final Map<String, Set<ContributorDescriptor>> _contributors = new HashMap<>();
-
   void registerFactory( @Nonnull final FactoryDescriptor factory )
   {
     _factories.put( factory.getElement().getQualifiedName().toString(), factory );
@@ -133,60 +119,11 @@ final class Registry
     return _injectors;
   }
 
-  /**
-   * Register the auto-fragment in the local cache.
-   *
-   * @param autoFragment the auto-fragment.
-   */
-  void registerAutoFragment( @Nonnull final AutoFragmentDescriptor autoFragment )
-  {
-    final String key = autoFragment.getKey();
-    _autoFragments.put( key, autoFragment );
-    autoFragment.markAsModified();
-    autoFragment.getContributors().addAll( getContributorsByKey( key ) );
-  }
-
-  @Nullable
-  AutoFragmentDescriptor findAutoFragmentByKey( @Nonnull final String key )
-  {
-    return _autoFragments.get( key );
-  }
-
-  @Nonnull
-  Collection<AutoFragmentDescriptor> getAutoFragments()
-  {
-    return _autoFragments.values();
-  }
-
-  /**
-   * Register the contributor in the local cache.
-   *
-   * @param contributor the contributor contributing to the auto-fragment.
-   */
-  void registerContributor( @Nonnull final ContributorDescriptor contributor )
-  {
-    _contributors.computeIfAbsent( contributor.getKey(), k -> new HashSet<>() ).add( contributor );
-  }
-
-  @Nonnull
-  Set<ContributorDescriptor> getContributorsByKey( @Nonnull final String key )
-  {
-    return _contributors.getOrDefault( key, Collections.emptySet() );
-  }
-
-  @Nonnull
-  Set<String> getContributorKeys()
-  {
-    return _contributors.keySet();
-  }
-
   void clear()
   {
     _factories.clear();
     _injectables.clear();
     _fragments.clear();
     _injectors.clear();
-    _autoFragments.clear();
-    _contributors.clear();
   }
 }