+++ processor/src/test/fixtures/bad_input/com/example/autofragment/ContainsFieldModel.java
@@ -1,9 +1,0 @@
-package com.example.autofragment;
-
-import sting.AutoFragment;
-
-@AutoFragment( "ContainsFieldModel" )
-public interface ContainsFieldModel
-{
-  int FOO = 1;
-}