+++ processor/src/test/fixtures/bad_input/com/example/autofragment/ContainsMethodModel.java
@@ -1,9 +1,0 @@
-package com.example.autofragment;
-
-import sting.AutoFragment;
-
-@AutoFragment( "ContainsMethodModel" )
-public interface ContainsMethodModel
-{
-  void doStuff();
-}