+++ processor/src/test/fixtures/bad_input/com/example/autofragment/EnumModel.java
@@ -1,8 +1,0 @@
-package com.example.autofragment;
-
-import sting.AutoFragment;
-
-@AutoFragment( "MyAutoFragment" )
-public enum EnumModel
-{
-}