+++ processor/src/test/fixtures/bad_input/com/example/autofragment/ExtendsSuperInterfaceModel.java
@@ -1,9 +1,0 @@
-package com.example.autofragment;
-
-import sting.AutoFragment;
-
-@AutoFragment( "ExtendsSuperInterfaceModel" )
-public interface ExtendsSuperInterfaceModel
-  extends Runnable
-{
-}