+++ processor/src/test/fixtures/bad_input/com/example/autofragment/ZeroContributorsModel.java
@@ -1,8 +1,0 @@
-package com.example.autofragment;
-
-import sting.AutoFragment;
-
-@AutoFragment( "ContainsFieldModel" )
-public interface ZeroContributorsModel
-{
-}