+++ processor/src/test/fixtures/bad_input/com/example/autofragment/duplicate_key/MyAutoFragment2Model.java
@@ -1,8 +1,0 @@
-package com.example.autofragment.duplicate_key;
-
-import sting.AutoFragment;
-
-@AutoFragment( "MyAutoFragment" )
-public interface MyAutoFragment2Model
-{
-}