+++ processor/src/test/fixtures/bad_input/com/example/contribute_to/NonCandidateTypeModel.java
@@ -1,8 +1,0 @@
-package com.example.contribute_to;
-
-import sting.ContributeTo;
-
-@ContributeTo( "Foo" )
-public class NonCandidateTypeModel
-{
-}