+++ processor/src/test/fixtures/bad_input/com/example/multiround/autofragment/already_generated/MyAutoFragment.java
@@ -1,8 +1,0 @@
-package com.example.multiround.autofragment.already_generated;
-
-import sting.AutoFragment;
-
-@AutoFragment( "MyAutoFragment" )
-public interface MyAutoFragment
-{
-}