+++ processor/src/test/fixtures/bad_input/com/example/multiround/autofragment/already_generated/MyFrameworkModel.java
@@ -1,6 +1,0 @@
-package com.example.multiround.autofragment.already_generated;
-
-@MyFramework
-class MyFrameworkModel
-{
-}