+++ processor/src/test/fixtures/bad_input/com/example/multiround/autofragment/already_generated/MyFrameworkModel2Impl.java
@@ -1,9 +1,0 @@
-package com.example.multiround.autofragment.already_generated;
-
-import sting.Injectable;
-
-@Injectable
-class MyFrameworkModel2Impl
-  extends MyFrameworkModel2
-{
-}