+++ processor/src/test/fixtures/input/com/example/autofragment/BasicAutoFragmentFrameworkModel.java
@@ -1,9 +1,0 @@
-package com.example.autofragment;
-
-import sting.ContributeTo;
-
-@ContributeTo( "BasicAutoFragmentModel" )
-@MyFrameworkComponent
-class BasicAutoFragmentFrameworkModel
-{
-}