+++ processor/src/test/fixtures/input/com/example/autofragment/BasicAutoFragmentModel.java
@@ -1,8 +1,0 @@
-package com.example.autofragment;
-
-import sting.AutoFragment;
-
-@AutoFragment( "BasicAutoFragmentModel" )
-public interface BasicAutoFragmentModel
-{
-}