+++ processor/src/test/fixtures/input/com/example/multiround/autofragment/MyAutoFragment.java
@@ -1,8 +1,0 @@
-package com.example.multiround.autofragment;
-
-import sting.AutoFragment;
-
-@AutoFragment( "MyAutoFragment" )
-public interface MyAutoFragment
-{
-}