+++ processor/src/test/fixtures/input/com/example/multiround/autofragment/MyFrameworkModel.java
@@ -1,6 +1,0 @@
-package com.example.multiround.autofragment;
-
-@MyFramework
-class MyFrameworkModel
-{
-}